requirements.txt
azure-functions
azure-storage-blob
requests
pandas
numpy
scikit-learn
joblib
